requirements.txt
python-telegram-bot==13.7
schedule==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0